Checkpoint engine health movement and Building state ownership
Preserve the incomplete migration as a reviewable draft. Native evidence is recorded in tools/spatial_oracle corpora and docs/research/TRACK_PROCESS_REPLAY_REGRESSION_NOTES.md; current validation and remaining failures are documented in the engine authority audit.

+++ src/map/rmg/emit.rs
@@ -153,7 +153,7 @@ pub fn populate(
         map_file.entities.push(MapEntity {
             owner: NEUTRAL_OWNER.to_string(),
             type_id: name.clone(),
-            health: NEUTRAL_HEALTH,
+            health: i32::from(NEUTRAL_HEALTH),
             cell_x: *x as u16,
             cell_y: *y as u16,
             facing: 0,
